Separate functionality testing into separate files.
Split PROPKA and ligand testing into separate files from basic tests.
