+++ pdb2pqr/tests/data/anthracene.mol2
@@ -1,0 +1,56 @@
+#       
+#       
+#       This file was generated by PRODRG version AA180301.0717
+#       PRODRG written/copyrighted by Daan van Aalten
+#       and Alexander Schuettelkopf
+#       
+#       Questions/comments to dava@davapc1.bioch.dundee.ac.uk
+#       
+#       When using this software in a publication, cite:
+#       A. W. Schuettelkopf and D. M. F. van Aalten (2004).
+#       PRODRG - a tool for high-throughput crystallography
+#       of protein-ligand complexes.
+#       Acta Crystallogr. D60, 1355--1363.
+#       
+#       
+@<TRIPOS>MOLECULE
+DRG  
+   14   16    1
+SMALL
+USER_CHARGES
+
+PRODRG MOLECULE
+@<TRIPOS>ATOM
+     1  CAB        -1.930     6.450    -2.470   C.ar      1 DRG       0.000
+     2  CAC        -2.360     7.390    -1.500   C.ar      1 DRG       0.000
+     3  CAD        -2.850     8.640    -1.920   C.ar      1 DRG       0.000
+     4  CAE        -3.300     9.550    -0.940   C.ar      1 DRG       0.000
+     5  CAF        -3.830    10.790    -1.350   C.ar      1 DRG       0.000
+     6  CAG        -4.300    11.700    -0.380   C.ar      1 DRG       0.000
+     7  CAN        -4.240    11.370     0.990   C.ar      1 DRG       0.000
+     8  CAM        -3.710    10.130     1.410   C.ar      1 DRG       0.000
+     9  CAL        -3.240     9.220     0.440   C.ar      1 DRG       0.000
+    10  CAK        -2.740     7.970     0.860   C.ar      1 DRG       0.000
+    11  CAJ        -2.300     7.060    -0.110   C.ar      1 DRG       0.000
+    12  CAI        -1.820     5.790     0.300   C.ar      1 DRG       0.000
+    13  CAH        -1.400     4.860    -0.680   C.ar      1 DRG       0.000
+    14  CAA        -1.460     5.190    -2.050   C.ar      1 DRG       0.000
+@<TRIPOS>BOND
+    1    2    1 ar
+    2    1   14 ar
+    3    2    3 ar
+    4    2   11 ar
+    5    4    3 ar
+    6    4    5 ar
+    7    4    9 ar
+    8    5    6 ar
+    9    6    7 ar
+   10    7    8 ar
+   11    9    8 ar
+   12    9   10 ar
+   13   11   10 ar
+   14   11   12 ar
+   15   12   13 ar
+   16   13   14 ar
+@<TRIPOS>SUBSTRUCTURE
+  1  DRG    1
