+++ pdb2pqr/tests/data/cyclohexane.mol2
@@ -1,0 +1,38 @@
+#       
+#       
+#       This file was generated by PRODRG version AA180301.0717
+#       PRODRG written/copyrighted by Daan van Aalten
+#       and Alexander Schuettelkopf
+#       
+#       Questions/comments to dava@davapc1.bioch.dundee.ac.uk
+#       
+#       When using this software in a publication, cite:
+#       A. W. Schuettelkopf and D. M. F. van Aalten (2004).
+#       PRODRG - a tool for high-throughput crystallography
+#       of protein-ligand complexes.
+#       Acta Crystallogr. D60, 1355--1363.
+#       
+#       
+@<TRIPOS>MOLECULE
+DRG  
+    6    6    1
+SMALL
+USER_CHARGES
+
+PRODRG MOLECULE
+@<TRIPOS>ATOM
+     1  CAA        11.290    18.580    21.950   C.3       1 DRG       0.000
+     2  CAB        11.410    19.880    22.760   C.3       1 DRG       0.000
+     3  CAC        11.080    21.080    21.870   C.3       1 DRG       0.000
+     4  CAF        12.020    21.110    20.660   C.3       1 DRG       0.000
+     5  CAE        11.900    19.810    19.860   C.3       1 DRG       0.000
+     6  CAD        12.230    18.610    20.750   C.3       1 DRG       0.000
+@<TRIPOS>BOND
+    1    1    2 1 
+    2    1    6 1 
+    3    2    3 1 
+    4    3    4 1 
+    5    4    5 1 
+    6    5    6 1 
+@<TRIPOS>SUBSTRUCTURE
+  1  DRG    1
