+++ pdb2pqr/tests/data/ethanol.mol2
@@ -1,0 +1,33 @@
+#       
+#       
+#       This file was generated by PRODRG version AA180301.0717
+#       PRODRG written/copyrighted by Daan van Aalten
+#       and Alexander Schuettelkopf
+#       
+#       Questions/comments to dava@davapc1.bioch.dundee.ac.uk
+#       
+#       When using this software in a publication, cite:
+#       A. W. Schuettelkopf and D. M. F. van Aalten (2004).
+#       PRODRG - a tool for high-throughput crystallography
+#       of protein-ligand complexes.
+#       Acta Crystallogr. D60, 1355--1363.
+#       
+#       
+@<TRIPOS>MOLECULE
+DRG  
+    4    3    1
+SMALL
+USER_CHARGES
+
+PRODRG MOLECULE
+@<TRIPOS>ATOM
+     1  CAA        21.780    -4.220    -8.790   C.3       1 DRG       0.000
+     2  CAB        20.740    -4.840    -7.860   C.3       1 DRG       0.150
+     3  OAC        20.850    -6.270    -7.890   O.3       1 DRG      -0.548
+     4  HAF        20.170    -6.670    -7.280   H         1 DRG       0.398
+@<TRIPOS>BOND
+    1    2    1 1 
+    2    2    3 1 
+    3    3    4 1 
+@<TRIPOS>SUBSTRUCTURE
+  1  DRG    1
