+++ pdb2pqr/tests/data/glycerol.mol2
@@ -1,0 +1,43 @@
+#       
+#       
+#       This file was generated by PRODRG version AA180301.0717
+#       PRODRG written/copyrighted by Daan van Aalten
+#       and Alexander Schuettelkopf
+#       
+#       Questions/comments to dava@davapc1.bioch.dundee.ac.uk
+#       
+#       When using this software in a publication, cite:
+#       A. W. Schuettelkopf and D. M. F. van Aalten (2004).
+#       PRODRG - a tool for high-throughput crystallography
+#       of protein-ligand complexes.
+#       Acta Crystallogr. D60, 1355--1363.
+#       
+#       
+@<TRIPOS>MOLECULE
+DRG  
+    9    8    1
+SMALL
+USER_CHARGES
+
+PRODRG MOLECULE
+@<TRIPOS>ATOM
+     1  OAD        -3.760     9.120     5.400   O.3       1 DRG      -0.548
+     2  HAF        -4.620     9.360     4.950   H         1 DRG       0.398
+     3  CAA        -2.870     8.530     4.430   C.3       1 DRG       0.150
+     4  CAB        -1.830     7.570     5.040   C.3       1 DRG       0.150
+     5  OAE        -2.410     6.360     5.540   O.3       1 DRG      -0.548
+     6  HAG        -1.690     5.750     5.860   H         1 DRG       0.398
+     7  CAC        -0.990     8.280     6.120   C.3       1 DRG       0.150
+     8  OAF        -0.110     9.260     5.560   O.3       1 DRG      -0.548
+     9  HAH         0.040     9.980     6.230   H         1 DRG       0.398
+@<TRIPOS>BOND
+    1    1    2 1 
+    2    3    1 1 
+    3    4    3 1 
+    4    4    5 1 
+    5    4    7 1 
+    6    5    6 1 
+    7    7    8 1 
+    8    8    9 1 
+@<TRIPOS>SUBSTRUCTURE
+  1  DRG    1
