+++ pdb2pqr/tests/data/naphthalene.mol2
@@ -1,0 +1,47 @@
+#       
+#       
+#       This file was generated by PRODRG version AA180301.0717
+#       PRODRG written/copyrighted by Daan van Aalten
+#       and Alexander Schuettelkopf
+#       
+#       Questions/comments to dava@davapc1.bioch.dundee.ac.uk
+#       
+#       When using this software in a publication, cite:
+#       A. W. Schuettelkopf and D. M. F. van Aalten (2004).
+#       PRODRG - a tool for high-throughput crystallography
+#       of protein-ligand complexes.
+#       Acta Crystallogr. D60, 1355--1363.
+#       
+#       
+@<TRIPOS>MOLECULE
+DRG  
+   10   11    1
+SMALL
+USER_CHARGES
+
+PRODRG MOLECULE
+@<TRIPOS>ATOM
+     1  CAB        -3.790    -3.860     0.280   C.ar      1 DRG       0.000
+     2  CAC        -4.440    -4.930    -0.380   C.ar      1 DRG       0.000
+     3  CAD        -5.830    -4.850    -0.630   C.ar      1 DRG       0.000
+     4  CAE        -6.470    -5.920    -1.280   C.ar      1 DRG       0.000
+     5  CAJ        -5.740    -7.060    -1.680   C.ar      1 DRG       0.000
+     6  CAI        -4.360    -7.150    -1.440   C.ar      1 DRG       0.000
+     7  CAH        -3.700    -6.080    -0.790   C.ar      1 DRG       0.000
+     8  CAG        -2.320    -6.160    -0.530   C.ar      1 DRG       0.000
+     9  CAF        -1.680    -5.100     0.130   C.ar      1 DRG       0.000
+    10  CAA        -2.400    -3.960     0.530   C.ar      1 DRG       0.000
+@<TRIPOS>BOND
+    1    2    1 ar
+    2    1   10 ar
+    3    2    3 ar
+    4    2    7 ar
+    5    3    4 ar
+    6    4    5 ar
+    7    5    6 ar
+    8    7    6 ar
+    9    7    8 ar
+   10    8    9 ar
+   11    9   10 ar
+@<TRIPOS>SUBSTRUCTURE
+  1  DRG    1
